+++ packages/frontend/src/constants/index.ts
@@ -1,3 +1,6 @@
 export const SNACK_BAR_ID = 'core-snack-bar'
 
 export const FEEDBACK_ID = 'core-feedback'
+
+export const OG_IMAGE_URL =
+  'https://www.twreporter.org/images/lawmaker/og_image.png'